requirements.txt
PySide6>=6.6
openai-whisper>=20231117
ffmpeg-python>=0.2.0
